+++ src/main/java/org/tomcatlogwatcher/dataaccess/DBConnector.java
@@ -12,22 +12,36 @@ public class DBConnector {
     private static final String UNAME = "sa";
     private static final String PASS = "";
 
-    public static Connection getConnection() throws Exception {
-        Class.forName(DB_DRIVER);
-        return DriverManager.getConnection(JDBC_URL, UNAME, PASS);
+    public static void initDBConnector() throws Exception {
+        try {
+            Class.forName(DB_DRIVER);
+        } catch (Throwable e) {
+            AppLogger.logFatal("DBConnector.initDBConnector: ", e);
+            throw e;
+        }
     }
 
-    public static void closeConnection(Connection conn) {
+    public static Connection getConnection() throws Exception {
+        try {
+            return DriverManager.getConnection(JDBC_URL, UNAME, PASS);
+        } catch (Throwable e) {
+            AppLogger.logFatal("DBConnector.getConnection: ", e);
+            throw e;
+        }
+    }
+
+    public static void closeConnection(Connection conn) throws Exception {
         try {
             if (conn != null && !conn.isClosed()) {
                 conn.close();
             }
-        } catch (Exception e) {
-            AppLogger.logSevere("Error in DBConnector.closeConnection()", e);
+        } catch (Throwable e) {
+            AppLogger.logFatal("DBConnector.getConnection: ", e);
+            throw e;
         }
     }
 
-    public static void closeDbObject(AutoCloseable... closeableObjs) {
+    public static void closeDbObject(AutoCloseable... closeableObjs) throws Exception {
         try {
             for(AutoCloseable closeable : closeableObjs) {
                 if (closeable != null) {
@@ -38,17 +52,19 @@ public class DBConnector {
                     }
                 }
             }
-        } catch (Exception e) {
-            AppLogger.logSevere("Error in DBConnector.closeConnection()", e);
+        } catch (Throwable e) {
+            AppLogger.logFatal("Error in DBConnector.closeConnection()", e);
+            throw e;
         }
     }
 
 
-    public static void rollback(Connection conn) {
+    public static void rollback(Connection conn) throws Exception {
         try {
             if (conn != null && !conn.isClosed()) conn.rollback();
-        } catch (Exception e) {
-            AppLogger.logSevere("Error in DBConnector.rollback()", e);
+        } catch (Throwable e) {
+            AppLogger.logFatal("Error in DBConnector.rollback()", e);
+            throw e;
         }
     }
 }